add ParallelTests using the container testtrait
